# src/OpenMetrc.Scraper/OpenApi/metrc-swagger-v2.yaml
swagger: 2.0
info:
  title: METRC API
  version: v2
paths:
...
  /sales/v2/receipts/{id}:
    delete:
      tags:
      - Sales
      summary: ''
      description: '<b>Permissions Required</b>: Sales'
      externalDocs:
        x-doc-key: delete_sales_v2_receipts_{id}
      operationId: DeleteSaleReceiptById
      parameters:
      - type: integer
        name: id
        in: path
        required: true
        description: Route parameter id
        format: int64
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The license number of the facility for which to archive receipt.
        format: ''
      responses:
        200:
          description: Ok
    get:
      tags:
      - Sales
      summary: ''
      description: '<b>Permissions Required</b>: Sales'
      externalDocs:
        x-doc-key: get_sales_v2_receipts_{id}
      operationId: GetSaleReceiptById
      parameters:
      - type: integer
        name: id
        in: path
        required: true
        description: Route parameter id
        format: int64
      - type: string
        name: licenseNumber
        in: query
        description: If specified, the Sales Receipt will be validated against the specified License Number. If not specified, the Sales Receipt will be validated against all of the User's current Facilities. Please note that if the Sales Receipt is not valid for the specified License Number, a 401 Unauthorized status will be returned.
        format: ''
      responses:
        200:
          description: Ok
...
  /sales/v2/paymenttypes:
    get:
      tags:
      - Sales
      summary: ''
      description: '<b>Permissions Required</b>: Sales Delivery'
      externalDocs:
        x-doc-key: get_sales_v2_paymenttypes
      operationId: GetSalePaymentTypes
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The license number of the facility for which to return the list of payment types.
        format: ''
      responses:
        200:
          description: Ok
  /sales/v2/receipts/active:
    get:
      tags:
      - Sales
      summary: ''
      description: '<b>Permissions Required</b>: Sales'
      externalDocs:
        x-doc-key: get_sales_v2_receipts_active
      operationId: GetSaleReceiptActive
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The license number of the facility for which to return the list of active receipts.
        format: ''
      - type: integer
        name: pageNumber
        in: query
        description: The number of the data page from which to return data.
        format: ''
      - type: integer
        name: pageSize
        in: query
        description: The number of records to return per page. Pagination is currently disabled by default. You can enable pagination on this query by specifying a value that does not exceed 20.
        format: ''
      - type: string
        name: salesDateStart
        in: query
        description: The sales date start timestamp. If specified, also specifying any of the last modified parameters will result in an error.
        format: date-time
      - type: string
        name: salesDateEnd
        in: query
        description: The sales date end timestamp. If specified, also specifying any of the last modified parameters will result in an error.
        format: date-time
      - type: string
        name: lastModifiedStart
        in: query
        description: The last modified start timestamp. If specified, also specifying any of the sales date parameters will result in an error.
        format: date-time
      - type: string
        name: lastModifiedEnd
        in: query
        description: The last modified end timestamp. If specified, also specifying any of the sales date parameters will result in an error.
        format: date-time
      responses:
        200:
          description: Ok
  /sales/v2/receipts/inactive:
    get:
      tags:
      - Sales
      summary: ''
      description: '<b>Permissions Required</b>: Sales'
      externalDocs:
        x-doc-key: get_sales_v2_receipts_inactive
      operationId: GetSaleReceiptInactive
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The license number of the facility for which to return the list of inactive receipts.
        format: ''
      - type: integer
        name: pageNumber
        in: query
        description: The number of the data page from which to return data.
        format: ''
      - type: integer
        name: pageSize
        in: query
        description: The number of records to return per page. Pagination is currently disabled by default. You can enable pagination on this query by specifying a value that does not exceed 20.
        format: ''
      - type: string
        name: salesDateStart
        in: query
        description: The sales date start timestamp. If specified, also specifying any of the last modified parameters will result in an error.
        format: date-time
      - type: string
        name: salesDateEnd
        in: query
        description: The sales date end timestamp. If specified, also specifying any of the last modified parameters will result in an error.
        format: date-time
      - type: string
        name: lastModifiedStart
        in: query
        description: The last modified start timestamp. If specified, also specifying any of the sales date parameters will result in an error.
        format: date-time
      - type: string
        name: lastModifiedEnd
        in: query
        description: The last modified end timestamp. If specified, also specifying any of the sales date parameters will result in an error.
        format: date-time
      responses:
        200:
          description: Ok
...
  /sales/v2/receipts:
    post:
      tags:
      - Sales
      summary: ''
      description: 'Please note: The SalesDateTime field must be the actual date and time of the transaction without the time zone. This date/time must already be in the same time zone as the Facility recording the sales. For example, if the Facility is in Pacific Time, then this time must be in Pacific Standard (or Daylight Savings) Time and not in UTC.</br><b>Permissions Required</b>: Sales'
      externalDocs:
        x-doc-key: post_sales_v2_receipts
      operationId: PostSaleReceipts
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The license number of the facility for which to record receipts.
        format: ''
      responses:
        200:
          description: Ok
    put:
      tags:
      - Sales
      summary: ''
      description: 'Please note: The SalesDateTime field must be the actual date and time of the transaction without the time zone. This date/time must already be in the same time zone as the Facility recording the sales. For example, if the Facility is in Pacific Time, then this time must be in Pacific Standard (or Daylight Savings) Time and not in UTC.</br><b>Permissions Required</b>: Sales'
      externalDocs:
        x-doc-key: put_sales_v2_receipts
      operationId: PutSaleReceipts
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The license number of the facility for which to update receipts.
        format: ''
      responses:
        200:
          description: Ok
...
  /sales/v2/receipts/finalize:
    put:
      tags:
      - Sales
      summary: ''
      description: '<b>Permissions Required</b>: Sales'
      externalDocs:
        x-doc-key: put_sales_v2_receipts_finalize
      operationId: PutSaleReceiptFinalize
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The license number of the facility for which to update finalized receipts.
        format: ''
      responses:
        200:
          description: Ok
  /sales/v2/receipts/unfinalize:
    put:
      tags:
      - Sales
      summary: ''
      description: '<b>Permissions Required</b>: Sales'
      externalDocs:
        x-doc-key: put_sales_v2_receipts_unfinalize
      operationId: PutSaleReceiptUnfinalize
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The license number of the facility for which to update unfinalized receipts.
        format: ''
      responses:
        200:
          description: Ok
...